remove(folder-icon-config): moved options to other folders
