+++ src/components/control-panels/folder-icon-shadow.jsx
@@ -3,15 +3,21 @@
 import Dropdown from "../inputs/dropdown";
 import Range from "../inputs/range";
 import Color from "../inputs/color";
+import Checkbox from "../inputs/checkbox";
 
 import { folderConfigStore } from "@/stores/folder-config";
 import { useTranslations } from "next-intl";
 
 export default function FolderIconOffset() {
+  const iconShadow = folderConfigStore((state) => state.iconShadow || false);
+  const iconType = folderConfigStore((state) => state.iconType);
+
   const setShadowColor = folderConfigStore((state) => state.setShadowColor);
   const setShadowBlur = folderConfigStore((state) => state.setShadowBlur);
   const setShadowOffset = folderConfigStore((state) => state.setShadowOffset);
   const setShadowOpacity = folderConfigStore((state) => state.setShadowOpacity);
+
+  const setIconShadow = folderConfigStore((state) => state.setIconShadow);
 
   const handleXChange = (event) => {
     const offsetX = event.target.value;
@@ -27,53 +33,65 @@ export default function FolderIconOffset() {
 
   const t = useTranslations("panelTitles");
   const tc = useTranslations("iconShadow");
+  const tcc = useTranslations("iconConfig");
 
-  if (!folderConfigStore((state) => state.iconShadow)) return null;
+  if (iconType === "none") return null;
 
   return (
     <Dropdown name={t("shadow")} icon="Blend">
-      <Color
-        label={tc("color")}
-        name="shadow-color"
-        defaultColor={"#000000"}
-        onChange={(e) => setShadowColor(e.target.value)}
+      <Checkbox
+        name="icon-shadow"
+        label={tcc("shadow")}
+        id="icon-shadow"
+        onChange={() => setIconShadow(!iconShadow)}
+        checked={iconShadow}
       />
-      <Range
-        label={tc("opacity")}
-        name="shadow-opacity"
-        id="shadow-opacity"
-        defaultValue={15}
-        min={0}
-        max={100}
-        onChange={(e) => setShadowOpacity(e.target.value)}
-      />
-      <Range
-        label={tc("blur")}
-        name="shadow-blur"
-        id="shadow-blur"
-        defaultValue={10}
-        min={0}
-        max={50}
-        onChange={(e) => setShadowBlur(e.target.value)}
-      />
-      <Range
-        label={tc("x")}
-        name="shadow-offset-x"
-        id="shadow-offset-x"
-        defaultValue={0}
-        min={-100}
-        max={100}
-        onChange={handleXChange}
-      />
-      <Range
-        label={tc("y")}
-        name="shadow-offset-y"
-        id="shadow-offset-y"
-        defaultValue={0}
-        min={-100}
-        max={100}
-        onChange={handleYChange}
-      />
+      {iconShadow && (
+        <>
+          <Color
+            label={tc("color")}
+            name="shadow-color"
+            defaultColor={"#000000"}
+            onChange={(e) => setShadowColor(e.target.value)}
+          />
+          <Range
+            label={tc("opacity")}
+            name="shadow-opacity"
+            id="shadow-opacity"
+            defaultValue={15}
+            min={0}
+            max={100}
+            onChange={(e) => setShadowOpacity(e.target.value)}
+          />
+          <Range
+            label={tc("blur")}
+            name="shadow-blur"
+            id="shadow-blur"
+            defaultValue={10}
+            min={0}
+            max={50}
+            onChange={(e) => setShadowBlur(e.target.value)}
+          />
+          <Range
+            label={tc("x")}
+            name="shadow-offset-x"
+            id="shadow-offset-x"
+            defaultValue={0}
+            min={-100}
+            max={100}
+            onChange={handleXChange}
+          />
+          <Range
+            label={tc("y")}
+            name="shadow-offset-y"
+            id="shadow-offset-y"
+            defaultValue={0}
+            min={-100}
+            max={100}
+            onChange={handleYChange}
+          />
+        </>
+      )}
     </Dropdown>
   );
 }
